requirements.txt
# annotated-types==0.7.0
# anyio==4.8.0
# appnope==0.1.4
# asttokens==3.0.0
# av==13.1.0
# beautifulsoup4==4.13.3
# blinker==1.9.0
# certifi==2025.1.31
# charset-normalizer==3.4.1
# click==8.1.8
# click-default-group==1.2.4
# cloup==3.0.5
# comm==0.2.2
# Cython==3.0.12
# debugpy==1.8.12
# decorator==5.1.1
# distro==1.9.0
# executing==2.2.0
# Flask==3.1.0
# Flask-Cors==5.0.0
# Flask-Login==0.6.3
# Flask-SQLAlchemy==3.1.1
# glcontext==3.0.0
# h11==0.14.0
# httpcore==1.0.7
# httpx==0.28.1
# idna==3.10
# ipykernel==6.29.5
# ipython==8.32.0
# isosurfaces==0.1.2
# itsdangerous==2.2.0
# jedi==0.19.2
# Jinja2==3.1.5
# jiter==0.8.2
# jupyter_client==8.6.3
# jupyter_core==5.7.2
# lxml==5.3.1
# manim==0.19.0
# manim-slides==5.4.2
# ManimPango==0.6.0
# mapbox_earcut==1.0.3
# markdown-it-py==3.0.0
# MarkupSafe==3.0.2
# matplotlib-inline==0.1.7
# mdurl==0.1.2
# moderngl==5.12.0
# moderngl-window==3.1.1
# nest-asyncio==1.6.0
# networkx==3.4.2
# numpy==2.2.3
# openai==1.63.2
# packaging==24.2
# parso==0.8.4
# pexpect==4.9.0
# pillow==11.1.0
# platformdirs==4.3.6
# prompt_toolkit==3.0.50
# psutil==7.0.0
# ptyprocess==0.7.0
# pure_eval==0.2.3
# pycairo==1.27.0
# pydantic==2.10.6
# pydantic-extra-types==2.10.2
# pydantic_core==2.27.2
# pydub==0.25.1
# pyglet==2.1.2
# pyglm==2.8.0
# Pygments==2.19.1

# python-dateutil==2.9.0.post0
# python-dotenv==1.0.1
# python-pptx==1.0.2
# pyzmq==26.2.1
# QtPy==2.4.3
# requests==2.32.3
# rich==13.9.4
# rtoml==0.12.0
# scipy==1.15.2
# screeninfo==0.8.1
# six==1.17.0
# skia-pathops==0.8.0.post2
# sniffio==1.3.1
# soupsieve==2.6
# SQLAlchemy==2.0.38
# srt==3.5.3
# stack-data==0.6.3
# svgelements==1.9.6
# tornado==6.4.2
# tqdm==4.67.1
# traitlets==5.14.3
# typing_extensions==4.12.2
# urllib3==2.3.0
# watchdog==6.0.0
# wcwidth==0.2.13
# Werkzeug==3.1.3
# XlsxWriter==3.2.2
# gunicorn
gtts
pymongo
cloudinary
manim==0.17.3
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
openai==1.12.0
requests==2.31.0
httpx==0.24.1
# werkzeug>=2.0.1
numpy>=1.21.0
scipy>=1.7.0
pillow>=8.3.1
# pycairo>=1.20.1